+++ packages/brick_offline_first/example/lib/app/models/mounty.dart
@@ -1,5 +1,4 @@
 import 'package:brick_offline_first/offline_first_with_rest.dart';
-import 'package:brick_offline_first_abstract/annotations.dart';
 
 import 'package:brick_offline_first_example/app/models/hat.dart';