+++ packages/brick_offline_first/lib/src/offline_first_with_rest/offline_first_with_rest_repository.dart
@@ -1,4 +1,3 @@
-import 'package:brick_core/core.dart';
 import 'package:brick_offline_first/src/offline_queue/request_sqlite_cache_manager.dart';
 import 'package:brick_sqlite/memory_cache_provider.dart';
 import 'package:brick_offline_first/offline_first.dart';
@@ -8,7 +7,6 @@ import 'package:brick_rest/rest.dart' show RestProvider, RestException;
 import 'package:brick_offline_first_abstract/abstract.dart' show OfflineFirstWithRestModel;
 import 'package:brick_sqlite_abstract/db.dart' show Migration;
 
-import 'package:brick_sqlite/sqlite.dart';
 import 'package:brick_offline_first/src/offline_queue/offline_queue_http_client.dart';
 import 'package:brick_offline_first/src/offline_queue/offline_request_queue.dart';