+++ packages/brick_offline_first/test/offline_first/offline_first_adapter_test.dart
@@ -1,5 +1,4 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:test/test.dart' show TypeMatcher;
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 
 import 'helpers/__mocks__.dart';