+++ packages/brick_sqlite_generators/test/sqlite_schema/test_all_field_types.dart
@@ -6,8 +6,6 @@ final output = r'''
 // GENERATED CODE DO NOT EDIT
 // This file should be version controlled
 import 'package:brick_sqlite_abstract/db.dart';
-// ignore: unused_import
-import 'package:brick_sqlite_abstract/db.dart' show Migratable;
 
 /// All intelligently-generated migrations from all `@Migratable` classes on disk
 final migrations = <Migration>{};