+++ packages/brick_offline_first_with_rest_build/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Remove unnecessary import in `brick.g.dart`
+
 ## 1.1.4
 
 * Always use `whereType<T>` casts after awaiting `Future.wait()` in Rest deserializing adapters.